# server/swagger.yml
openapi: "3.0.0"
info:
  title: "E-Commerce Platform API"
  version: "1.0.0"
  description: |
    # E-Commerce REST API Documentation
    
    Modern full-stack e-commerce API with Domain-Driven Design architecture.
    
    ## API Overview
    - **Base URL**: `/api/v1`
    - **Authentication**: Bearer token (JWT)
    - **Rate Limiting**: Applied to auth and payment endpoints
    - **Response Format**: JSON
    
    ## Domain Routes
    - **Authentication** (`/auth`) - User registration, login, token management
    - **Identity** (`/identity`) - User profiles, addresses, payment methods
    - **Catalog** (`/catalog`) - Products, categories, wishlist
    - **Ordering** (`/ordering`) - Cart, orders, checkout, payments, shipping
    - **Vendors** (`/vendors`) - Vendor management and applications
    
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://localhost:5000"
    description: "Development"
  - url: "https://staging.ecommerce.example"
    description: "Staging"
  - url: "https://api.ecommerce.example"
    description: "Production"
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      security: []
      operationId: registerUser
      summary: "Register a new user"
      description: "Create a new user account with email and password"
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: "User registered successfully"
        "400":
          description: "Validation error"
        "409":
          description: "Email already registered"
  /api/v1/auth/login:
    post:
      security: []
      operationId: loginUser
      summary: "Authenticate a user"
      description: "Login with email and password. Returns JWT access token and refresh token."
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: "Login successful"
        "401":
          description: "Invalid credentials"
  /api/v1/auth/refresh-token:
    post:
      security: []
      operationId: refreshToken
      summary: "Refresh access token"
      description: "Get a new access token using the refresh token"
      tags: ["Authentication"]
      responses:
        "200":
          description: "Token refreshed successfully"
        "401":
          description: "Invalid or expired refresh token"
...
  /api/v1/auth/verify-email:
    get:
      security: []
      operationId: verifyEmail
      summary: "Verify email address"
      description: "Verify user's email using token sent in registration email"
      tags: ["Authentication"]
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Email verified successfully"
        "400":
          description: "Invalid or expired token"
...
  /api/v1/auth/reset-password:
    post:
      security: []
      operationId: resetPassword
      summary: "Reset password"
      description: "Reset password using token from reset email"
      tags: ["Authentication"]
      responses:
        "200":
          description: "Password reset successfully"
        "400":
          description: "Invalid or expired token"
...
  /api/v1/catalog/wishlist:
    get:
      operationId: getWishlist
      summary: "Get user's wishlist"
      description: "Retrieve all products in the user's wishlist"
      tags: ["Catalog"]
      responses:
        "200":
          description: "Wishlist retrieved successfully"
        "401":
          description: "Unauthorized"
    post:
      operationId: addToWishlist
      summary: "Add product to wishlist"
      description: "Add a product to the user's wishlist"
      tags: ["Catalog"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: "Product added to wishlist"
        "401":
          description: "Unauthorized"
    delete:
      operationId: clearWishlist
      summary: "Clear entire wishlist"
      description: "Remove all products from the user's wishlist"
      tags: ["Catalog"]
      responses:
        "200":
          description: "Wishlist cleared successfully"
        "401":
          description: "Unauthorized"
  /api/v1/catalog/wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      summary: "Remove product from wishlist"
      description: "Remove a specific product from the user's wishlist"
      tags: ["Catalog"]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Product removed from wishlist"
        "401":
          description: "Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...